+++ tests/unit/pixel/image/inspectors/gd/PixelColorTest.php
@@ -1,0 +1,61 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\tests\unit\pixel\image\inspectors\gd;
+
+use mako\pixel\exceptions\PixelException;
+use mako\pixel\image\Color;
+use mako\pixel\image\Gd;
+use mako\pixel\image\inspectors\gd\PixelColor;
+use mako\pixel\image\operations\Point;
+use mako\tests\TestCase;
+use PHPUnit\Framework\Attributes\Group;
+use PHPUnit\Framework\Attributes\RequiresPhpExtension;
+
+#[Group('unit')]
+#[RequiresPhpExtension('imagick')]
+class PixelColorTest extends TestCase
+{
+	/**
+	 * Note that the colors extracted by GD aren't 100% accurate.
+	 */
+	public function testPixelColor(): void
+	{
+		$image = new Gd(__DIR__ . '/../../fixtures/001.png');
+
+		$color = $image->inspect(new PixelColor(new Point(0, 0)));
+
+		$this->assertInstanceOf(Color::class, $color);
+
+		$this->assertSame('#B01000', $color->toHexString());
+
+		$color = $image->inspect(new PixelColor(new Point(0, 100)));
+
+		$this->assertInstanceOf(Color::class, $color);
+
+		$this->assertSame('#0070C0', $color->toHexString());
+
+		$color = $image->inspect(new PixelColor(new Point(0, 275)));
+
+		$this->assertInstanceOf(Color::class, $color);
+
+		$this->assertSame('#007000', $color->toHexString());
+	}
+
+	/**
+	 *
+	 */
+	public function testPixelColorWithInvalidPosition(): void
+	{
+		$this->expectException(PixelException::class);
+		$this->expectExceptionMessageIs('Pixel coordinates [ 1000, 100 ] are outside image bounds [ 300, 300 ].');
+
+		$image = new Gd(__DIR__ . '/../../fixtures/001.png');
+
+		$image->inspect(new PixelColor(new Point(1000, 100)));
+	}
+}